kernel: move port.h/pci.h from x86_64 to shared code and refactor how we load PCI devices
