+++ src/kernel/core/port.c
@@ -1,4 +1,4 @@
-#include "port.h"
+#include <core/port.h>
 
 uint8_t port_byte_in(uint16_t port)
 {
